+++ src/main/java/com/deadline826/bedi/login/Controller/LoginController.java
@@ -17,7 +17,12 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.util.Random;
+
+import static com.deadline826.bedi.security.JwtConstants.*;
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 
 @RestController
@@ -39,7 +44,6 @@ public class LoginController {
         while (true){  //고유 아이디 생성
             userDto.setId(Long.parseLong(String.valueOf(1000000000 + random.nextInt(900000000))));
             User user = userService.findUserById(userDto.getId());
-            System.out.println("user = " + user);
             if (user==null)
                 break;
         }
@@ -53,29 +57,41 @@ public class LoginController {
 
 
     @PostMapping("/login")
-    public ResponseEntity<TokenDto> login(@RequestBody LoginDto loginDto)
-            throws AuthenticationNumberMismatchException {
-
-        // 이메일로 유저를 불러와
-        //User user = userService.findUserByEmail(loginDto.getEmail());
-
-
-//        // dto로 변환하고
-//        UserDto userDto = new UserDto();
-//        userDto.setEmail(user.getEmail());
-//        userDto.setUsername(user.getUsername());
-//        userDto.setPhone(user.getPhone());
-//        userDto.setPassword(loginDto.getPassword());
-//        userDto.setId(user.getId());
+    public ResponseEntity<TokenDto> login(@RequestBody LoginDto loginDto) throws AuthenticationNumberMismatchException {
 
         // redis 를 아용하여 로그인 정보를 1분간 임시저장한다 (PW 검사시 이용됨)
         verifyPasswordService.makeTempPasswordStorage(loginDto.getEmail(),loginDto.getPassword());
 
-        //변환된 dto로 로그인 진행
+        // loginDto 로 로그인 진행
         TokenDto loginToken = userService.login(loginDto);
 
 
         // 토큰정보를 반환한다
         return ResponseEntity.ok().body(loginToken);
+    }
+
+
+
+    //refreshToken 을 이용하여 accessToken 가져오기
+    @GetMapping("/refresh")
+    public ResponseEntity<TokenDto> refresh(HttpServletRequest request, HttpServletResponse response) {
+        String authorizationHeader = request.getHeader(AUTHORIZATION);
+        if (authorizationHeader == null || !authorizationHeader.startsWith(TOKEN_HEADER_PREFIX)) {
+            throw new RuntimeException("JWT Token이 존재하지 않습니다.");
+        }
+        String refreshToken = authorizationHeader.substring(TOKEN_HEADER_PREFIX.length());
+        TokenDto tokens = userService.refresh(refreshToken);  //refreshToken 을 넣으면 accessToken 이 반환 됨
+        response.setHeader(AT_HEADER, tokens.getAccessToken());   //위의 accessToken 을 헤더에 넣고
+        if (tokens.getRefreshToken() != null) {              // refreshToken 의 만료기간이 다가와서 새로 받은 refreshToken 이 있다면
+            response.setHeader(RT_HEADER, tokens.getRefreshToken());  // refreshToken 도 같이 헤더에 넣는다.
+        }
+        return ResponseEntity.ok(tokens);   //화면에 출력
+    }
+
+    // 내정보 가져오기
+    @GetMapping("/my")
+    public ResponseEntity<Long> my(HttpServletRequest request) {
+        User user = userService.getUserFromAccessToken();
+        return ResponseEntity.ok(user.getId());
     }
 }
